pyproject.toml
[project]
name = "dan-quotes-service"
version = "0.1.0"
description = "FastAPI backend + minimal web UI for submitting and moderating Dan quotes, backed by InstantDB."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi==0.115.6",
  "uvicorn[standard]==0.34.0",
  "jinja2==3.1.5",
  "python-dotenv==1.0.1",
  "python-multipart==0.0.12",
  "httpx==0.28.1",
  "pydantic==2.10.4",
  "PyJWT[crypto]==2.10.1",
  "motor==3.6.0",
  "pytest==8.3.4",
  "pytest-asyncio==0.24.0",
]
